[Lacewing][Suspend-Resume]Add an option to disable timer duration check
idle-suspend-auto-resume based PowerBAT tests are failing because of
b/355511475, until this issue is resolved, we are introducing an option
to disable the timer duration check. Please note that PowerBAT users are
still expected to have a test with timer duration check enabled as this
is how we found this bug in the first place. However users now have an
option to also run with timer duration check disable as a separate test

BUG: b/355511475

Change-Id: I89368a5605758626feb0fa1c14ac8f0aa37dfc02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121660
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Pavan Kumar Juturu <jpavankumar@google.com>
Reviewed-by: Harsha Priya N V <harshanv@google.com>