[sestarnix] Allow FsNode's in /proc/<pid> to be labeled based on a Task.
This CL turns the SecurityState of FsNode's into a container for an enum
that has 3 states:
* Uninitialized
* SecurityId
* WeakRef<Task>

This makes it possible for an FsNode to have an effective SecurityId
that tracks that of a Task, even if that Task's SecurityId changes.

This CL labels the /proc/<pid>/ directories, but not the children of
/proc/<pid>/.

Later CLs will add logic to propagate Task-based FsNode labeling to
child FsNode's as they are created.

Bug: 322850026

Change-Id: Id7873b7e62b23b5910efbc3d93acd35387fdfc90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103653
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Nelly Vouzoukidou <nellyv@google.com>