[build][sdk] Disallow use of "public" SDK category
There is no consumer of the "public" SDK category, and compatibility
enforcement is bypassed when using it, so disallow specifying this
category until it is needed and properly implemented.

Add checks to ensure that only valid and supported SDK categories are
used in atoms in the SDK. (These checks may not be executed if one of the
more specific `sdk_()` templates does not generate an `sdk_atom()`
for a target.) Also update the documentation.

Change a few existing instances of `[sdk_]category = "public"` to
"partner", remove `sdk_category from two FIDL libraries that aren't
intended to be in an SDK, update an example, and fix an invalid
category.

Bug: 366221510
Change-Id: Ic2fe95af7b2059e54467fb7f020c4e3e22522536
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120677
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>