[firmware][sdk] Replace SDK category "excluded" instances
Replace instances of "excluded" in uses of SDK templates for the
Firmware SDK. The Firmware SDK uses these templates and thus must
specify a category, but it is not a real IDK so the categories are
currently irrelevant. These are the last uses of "excluded", so
replacing them will allow us to eliminate this category.

Introduce a "firmware_sdk" pseudo category to `sdk_source_set()`, which
is the template used for almost all atoms that make up the Firmware SDK.
The template converts this to "internal" when constructing the
sdk_atom() but after performing steps specific to @internal_sdk.

"experimental", which is used by some such instances in third-party
repos is treated the same as "sdk_firmware" until those can be updated.

For documentation and the collection, "internal" is sufficient.

Bug: 365996258, 331991540
Change-Id: I5c73b97c2cbd7e3850b098cfefbd7afbfa733087
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120682
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: David Pursell <dpursell@google.com>