[fuzz_test_runner] Fix 'use_next_vdso: "true"' support
The fuzz runner can be utilized by certain fuzzers that require
the "next" vDSO (e.g., Starnix), but the "next" vDSO was not propagated to
the components launched by the runner.
fuzz_test_runner binary now clones the vDSO that it was started
with when launching a new component(e.g. libfuzzer_engine)

Bug: 364932357
Change-Id: I4a14e7458afcc0507159742b1cd8620b9fb9dba6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114215
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Dzmitry Yatsushkevich <dmitryya@google.com>