[fxfs] Ignore deleted extents during fsck.
Deleted extents for attributes on objects that have not been tombstoned
were being identified as orphaned extents. They get compacted away
during iteration if the object was tombstoned, but if not tombstoned
they are still present with the None value so we should ignore them.

Change-Id: Ifff0d70a369200848df19bf3f72d57172824734c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121383
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>