[ffx][fho] Adding external subtool manifest support
This adds building a JSON manifest of the external subtools built as
part of the build, and a configuration key to use to pass it into
ffx. When the subtool manifest is configured, the manifest is read to
configure the external ffx subtools. If the manifest is not present,
it scans the appropriate directories looking for subtools.

By adding the manifest processing, ffx can be used hermetically in
the build, and still access external subtools.

Bug: 366318546
Change-Id: I417a0da8075f94e092367ae6ab63fecba7dcfe91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121714
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>