[usb][dwc3] move uep lock inside of the endpoint
There is a 1:1 relationship between uep and underlying endpoint struct,
so this can be moved behind the endpoint struct. This is a necessary
step to add the Endpoint protocol impl, and it got messy doing it at
the same time. I refactored this change out for simplicity.

Test: Boot kola and ssh login over CDC
Bug: b/350802129
Change-Id: I0adf917774614a5232e8a666cc141b3bd179c9f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121455
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>