[sandbox] Router interface admits null request
Update the interface of Router to allow a request to be `None`. If the
request is `None`, the router may supply a default request (or it may
choose to return an error and require an explicit request).

This commit does not actually change the behavior of any router to
accept a null request. That will come in a followup.

Note that `debug` (not present in fidl) is still required, because we
wish to require the caller to choose whether they are performing a debug
route.

Bug: 366458428

Change-Id: I16ae8e8a7ed9b15dba7248cdb9661268c27053f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112031
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
API-Review: Gary Bressler <geb@google.com>