[zx][rs] Add API for boot time and timers.
Note that these APIs do not yet read from CLOCK_BOOT because it is
not yet stable. This is fine as long as the monotonic clock does
not yet pause during suspends, as the two timelines are currently
equivalent. The naming for these APIs is not yet final and may be
revised before the relevant RFCs are ratified.

This will allow clients to begin migrating to the new APIs in
advance of a future change which will switch these APIs to be
based on the new timeline.

Bug: 328306129
Change-Id: I4151c92aff11e62991fa46b4d44716acf08d15e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106092
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>