[network/tests] Test ping forwarding
Add a test verifying that a client on Network A can reach the router's
address on Network B, if the router is configured to forward between the
two networks.

Essentially, this verifies that the Netstack operates as a Weak Host
when configured as a router. Currently it passes against NS2 but fails
against NS3.

While here, refactor the original "forwarding" test so that the setup
can be shared.

Bug: 368343873
Change-Id: Idaeaabdd15d6cc3252663f77e8df41e5760253d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122153
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>