[pkg] Deduplicating RepoStorageType, ConflictMode
This removes the recently added pkg:: enums that mirror
the FIDL enums for the StorageType and Alias conflict mode. These
mirrored versions are used to add serialization to the enum.

There is still a mirror in fidl_ffx_developer_ext, but that can
be handled in a follow-up CL.

Bug: 367668384
Change-Id: Id10fb17a012a99bcc88dc6b9053d6a38dbab7e64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120573
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>