[ffx] Add target disconnect request
When a test causes the target to reboot, the daemon can be left
with a stale connection. While that connection will eventually
be detected, the detection is based on timeouts that are changeable.
If the test wants a predictable result, it needs to inform the daemon
that the connection it has is stale, i.e. that the daemon should drop
the existing connection.

By itself, this CL does nothing since there are no clients for the
new Target.Disconnect() FIDL request. But there should be a follow-up
change to add the method to Fuchsia Controller. And possible a
new ffx subtool, `ffx daemon disconnect`, to give a similar ability
to non-FC clients.

Bug: 367453247
Change-Id: I1671d349574e7dff4c7d14a6078705081058f106
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122213
Reviewed-by: Andrew Davies <awdavies@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>