[ffx] Add "ffx daemon disconnect"
Add a command to tell the daemon to disconnect from a target.
This is useful in a reboot test, which will know that the connection
the daemon is maintaining is no longer valid. It can be used to
be explicit, rather than relying on the unknown timing of the daemon
to detect that the target has gone away.

Bug: 367453247
Change-Id: If5d88076ca43fd6bc53befdcc002c41f667fce3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122219
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>