[fxfs] move has_overwrite_extents flag to attribute value
Instead of having it as part of the main object value, move the flag to
be per-attribute. This is primarily an optimization - when opening an
object we already look up the data attribute, so we can use that
existing check to find this flag as well.

Bug: 293943124
Change-Id: If78b43d379d3da4a55faedadf39539df89937e97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122293
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>