[bazel_sdk] add ability for bt-host to not include c++
The bt-host wants to statically link c++ to test its impact on size.
However, we don't currently have a way to not bundle libc++ libraries in
a package if it statically links in c++. This CL adds a hack which
allows them to remove the libraries from the package. When we build out
better logic for determining if libc++ should be included we will remove
this hack.

BUG: 368434718
Change-Id: I868dc2c146b54a00687308720716520d95b13253
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122452
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>