[bazel_sdk] do not package libcxx when statically linking
When a user specifies that they want to statically link c++ via the
static_cpp_standard_library feature we want to avoid packaging the
libcxx dist libraries. This adds that check and removes the dist targets
in the fuchsia_wrap_cc_binary macro.

In the future we will need to move this logic into the dist target
itself.

Additionally, this CL adds new tests to verify that a package that
contains a mix of statically linked libraries and binaries that are not
statically linked still include the dist files.

BUG: 368434718
Change-Id: Ifed1dec23e97a1a35c6e8a5b23ee17991b775e40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122625
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>