[kernel] Remove stale header dev/uart.h
This header defined the contract from dev/uart to the platform, which
was removed earlier. Unfortunately it was under the kernels global
include, which meant when dev/uart target was removed it was not flagged
by not being explicit about it's dependencies
(//zircon/kernel/dev/pdev/uart).

Bug: 364713557
Change-Id: I77d734738be1df8da5a92dc3f69c18790345d0fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119507
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>