[starnix][touch-power-policy] Correct device metadata
The metadata for this device is incorrect and harmful. We cannot grab
minor device 0 for the MISC major device. The device number is assigned
to the Logitech bus mouse.

Instead, we should register this starnix-specific device as a dynamic
device, which assigns it a device number that has been reserved for
unknown devices.

Also, because the device is no longer in the MISC major device, this CL
switches the registration over to the "starnix" class, which is also
likely to be incorrect, but it's the class we've been using for these
Starnix-specific devices for the time being.

Change-Id: I5aff509378c7c10ed0bce020b0c3d017b99b01f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122992
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>