[starnix] Split overloaded FsNode::socket field
This field had two jobs: For FDs created by syscalls like `socket()`, it
held the underlying socket so socket syscalls could access it through
the FD. For unix domain socket nodes, it held the socket bound to that
node. It turns out that for the first type of FD the socket is already
held by the FileOps, and nothing needs to access it through the FsNode.
So switch the first type of field accessors to go through the FileOps,
and rename the field to `bound_socket` for use for bound unix domain
sockets.

Change-Id: Ib917b76e33d5ec2f5266ca7fce585329786f407c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121712
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>