[kernel][userboot] Construct RoDso from a VmObject
This change refactors RoDso to be constructible from an explicitly
provided VMO, factoring out the "embedded" logic (which we'll soon
invalidate) and the kernel handle logic (which is not appropriate for
the userboot image case). This eases the upcoming removal of the
embeddings of userboot and the vDSO from the kernel.

Bug: 1097802
Change-Id: I4236562b28365bfc992584a6fa95b66d9121607b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121717
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>