[netdevice] Fix race on TX in-flight counter
Currently in Inner::send a buffer is queued to be sent to the driver
before the number of TX buffers in flight is incremented. This means
that it's possible for the buffer to be sent to the driver and complete
a round trip back to the client before the in-flight counter is
incremented, at which point it's decremented while currently at 0,
causing a debug assertion to trip.

Fix this race by incrementing the in flight buffer counter before
actually queueing it to be sent to the driver.

Fixed: 369180969, 368652664, 368689639
Multiply: tun_socket_benchmarks_with_netstack3_test.sh
Change-Id: I3a719232b735b59bd0762ea41a20b5670e5bf637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123853
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>