[cmdline] keep local console enabled with kernel.shell=true set
In the past, when setting kernel.shell=true the user shell was disabled
because the system continued to start user space and only one shell was
started (the kernel's). A while back it was changed such that enabling
the kernel shell stops the boot process before starting user space so no
need to disable the user shell.

Remove this particular interlock which now allows the user to drop into
the kernel shell, run some commands, and then exit the kernel shell and
continue a full user space with a user shell.

Change-Id: I3ea4b2f33b9f3c4c723c9c799a513513a826f7f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121458
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>