[kernel][target] remove the entire vestigial target/ directory
Forever ago this held the per target (soc/board/etc) layer of the
kernel, which was generally refactored into generic platforms when the
kernel was refactored to be per arch and not per board.

The dir was left around to hold the old boot shims, which have been
replaced with a new style boot shim.

Bug: 42147821

Change-Id: I5724a0940665f8a8b2c73732e9fe2bdedb81328f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121713
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>