[sdk][build] Enforce meta.stable in sdk_atom()
Require that `meta.stable` is specified for atom types that support
unstable atoms and that it is set to an appropriate value for the SDK
category. This requires adding `stable` to the `meta` in
`fidl_library()`. Ensure that `stable` is not passed to the static and
shared library templates as an additional precaution.

Similarly, ensure an `api` file is specified except when the atom is
unstable, including when it is in the SDK category "internal".

Also note that the value of `stable` in this build manifest does not
affect the IDK, though it should match the value in the IDK meta.json.

Bug: 366221510, 364727245
Change-Id: I781efe78bee82b7160bf42e8f5b36b808fcc74af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122231
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>