[sestarnix] Eagerly initialize FsNode labeling details
Previously labels for FsNodes were resolved on-demand, with the
fs_node_effective_sid() helper calling on to the
fs_node_resolve_security_label() helper.

Labels are now resolved for unlabeled FsNodes directly by the
fs_node_init_with_dentry() hook, when they are first linked into
a DirEntry. e.g. for genfscon they will be labeled based on their
FileSystem-relative path, and for fs_use_xattr their security
attribute will be read, to populate a label.

Special cases, such as creation of an FsNode for a new file, may
label the node prior to it being linked into a DirEntry for the
first time, in which case the fs_node_init_with_dentry() hook is
a no-op.

This requires that any FsNodes created in FileSystems prior to the
SELinux policy being loaded are tracked, and proactively labeled
along with the FileSystem. The initial implementation retains a
set of weak references to the relevant DirEntries, to allow node
labeling to take account of the FileSystem-relative path.

Bug: 366405530
Change-Id: I8f643971636346b3841512997f9efc2297076c04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119412
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>