[build] Move build_only_labels to top-level //:build_only group and use with subbuilds instead of universe
Universe is the wrong target to use for sub-builds, as it's
specifically a dependency of the product being compiled,
and sub-builds are really trying to do a build-only type of
operation.  We already have a build_only_labels mechanism,
although it's somewhat hidden.

This moves the build_only_labels mechanism into //BUILD.gn
and creates a top-level group for it to use that will play
nicely with the GN --root-pattern mechanism.

Change-Id: Idc980cd6060dd7fc68f48ec78c7d4633a7823a25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123882
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>