[netstack3] Update comment about packet metadata
It's not guaranteed that a packet will have its conntrack connection
stored in its metadata by the time it's at the last filtering hook in
its journey (LOCAL_INGRESS for locally-destined packets or EGRESS for
remotely-destined packets). However, comments currently indicate that
this is due to packet reassembly occurring after INGRESS, which is no
longer the case as of I0dfbad73b6fb4c86f5e46732701e27bd95977843. That
change did make it so that most packets will have a connection in their
metadata by the time they hit LOCAL_INGRESS or EGRESS, but there is are
still other scenarios when this may not be the case:

 * In multicast forwarding, when a packet is forwarded out of multiple
   interfaces, only the first forwarded packet will have a connection in
   its metadata.
 * Packets that are not tracked by conntrack will never have a
   connection in their metadata.

Update these comments to more correctly reflect this.

Change-Id: Ie2144d4dbbdde93288d2e8fe938144d0f98915df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121566
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>