[netstack3] Expect NAT configured before conn finalized
NAT should always be configured for a connection before it is inserted
in the connection tracking table. This did not use to be the case;
previously, incoming packets were reassembled in between the INGRESS and
LOCAL_INGRESS hooks, which meant the first packet in the flow did not
traverse the NAT routines on the INGRESS hook (which configure DNAT).

I0dfbad73b6fb4c86f5e46732701e27bd95977843 changed this behavior such
that incoming packets are always reassembled before they hit any
filtering hook; now, the first packet of any flow should have NAT
configured for it (whether or not that configuration is "no-op") before
its connection is finalized and inserted in the connection tracking
table.

Change-Id: I317c3af9bab2c32ccb98a0975c656638ebaae72f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121569
Reviewed-by: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>