[tee] Make app endpoint access more thread-safe
The last 'fix' preserved a race condition in the original connection
code:
(1) Locked: see if we have an existing connection to an app; if so
return it;
(2) Unlocked: Create new connection
(3) Locked: Insert new endpoint into container

Since (2) is unlocked, there's a race to create and insert the new app
endpoint, which can trigger a "not yet contained" assertion in the
insertion logic. We address this by replacing the more atomic get/insert
methods with a compound "connect, creating a new connection if one does
not yet exist" method.

Bug: 348473568
Change-Id: I51a743e803d81fd7784c4b834f80e23efb2817c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123930
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>