[dl] Persist a RuntimeModule's module tree
This CL will store a list of the breadth-first order module tree
(inclusive of the root module) in the RuntimeModule for the dlopen-ed
module. This list is currently passed to relocations as is, but in the
future global modules will be prepended to this list for relocations.

This list is used directly by dlsym for symbol lookup.

This CL does not yet set the module tree for a module that was loaded
as a dependency of another module, which will be resolved in a future
CL.

Many tests that were skipped with kDlSymSupportsDeps are now enabled on
libdl. The kDepModuleHasDepTree feature flag is introduced to skip test
scenarios that exercise the unsupported case.

Bug: 354786114
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I9fa821894a357f3697eff94d5d80ada6343d9b9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116692
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>