[dl] Persist a RuntimeModule's direct deps
This CL will store a list of references to only the direct dependencies
of the RuntimeModule. This list is populated in the first BFS traversal
of the sessions_modules_ list, when direct dependencies are enqueued.
This list contains references to new dep modules that were loaded with
the root, as well as previously loaded dep modules.

EnqueueModule will now return a pointer to the loaded or new
RuntimeModule that was created while iterating over the session_modules_
list so that it can be added to the parent module's direct_deps list.

This list is used in the second BFS traversal of the root
module's module tree. Iterating over this list replaces the previous
`enqueue_deps` algorithm that checked whether a dep module was new to
the session or previously loaded in the second loop.

Bug: 354786114
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: If4d94588ffe1965629e46d7fa33810fb1573703f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116532
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>