[logs] Create separate shard for SDK use
In order to migrate in-tree users to dictionaries,
we need to make breaking route changes.
Out-of-tree users can't use dictionaries yet,
and even when they are able to, we can't atomically
migrate them. A gradual migration of OOT users
will be required to route dictionaries everywhere.
To allow in-tree users to migrate to new routes,
create a separate shard to differentiate between
dictionaries users and non-dictionary users.

Bug: b/345827642
Change-Id: I99173d2a52f2547dc2536288e1c13b923a86b7e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121570
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>