[vsock] Move bind state to server
The fidl protocol was split such that there is a channel connection per
socket. This simplifies some state management. Additionally, Listen2 was
split into 2 separate calls, Bind on the Connector and Listen on the
Listener. This split allows the vsock_service component to maintain
socket state entirely on the server. This is helpful for the case in
which we need to duplicate or migrate the socket to another process.
Neither of these operations are yet supported, but they can be more
easily supported after this change.

Fixed: 368645901
Change-Id: Ib58f5cdbdb204085e1a2c70eb153a3c6a2b2b7e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123881
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>