[lacewing] Improve hermetic_exe portability.
Hermetic Lacewing executables can now be launched without additional considerations about the CWD - which will streamline the test execution UX for OOT consumers.

Fuchsia Controller's import hook requires the hermetic exe to be launched in a CWD that contains the shared libraries or a host_x64/ directory. This has been working fine in infra so far because there just so happens to be a host_x64/ dir in the CWD (//out/default).

Change-Id: I841f95f61e821e3e5f67dfa7fca4c6e4e4c39784
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124715
Commit-Queue: Kevin Cho <chok@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>