[cml] Revise syntax for dynamic dictionaries
Instead of:

```
capabilities: [
  {
    dictionary: "dict",
    extends: "program/svc/fuchsia.component.sandbox.Router",
  },
]
```

The new syntax is:

```
capabilities: [
  {
    dictionary: "dict",
    path: "/svc/fuchsia.component.sandbox.Router",
  },
]
```

This harmonizes better with other capability types. The semantics of
`extending from program` diverge from `extending from
parent/self/child`, and it's a good idea to reflect this in the API.

In a followup, I will land a change to decouple dynamic dictionaries
from the extension codepath entirely.

In addition, disallow `extends` and `offer to self/dict` in this mode -
they no longer make sense as in the followup we're no longer creating an
intermediate dictionary. If users want this behavior, they can always
create a second dictionary and extend the dynamic one.

Bug: 368155371
Change-Id: I836f3c6327d12a9ef5b521634b56a4434daf0f13
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121718
API-Review: David Gilhooley <dgilhooley@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Gary Bressler <geb@google.com>