[assembly] Move product.base_packages to bootfs for no_image
When the product indicates that a filesystem image (fvm or fxfs) should
not be generated:
  platform.storage.filesystem.image_mode = "no_image"

Then all base packages should be moved to bootfs. In this situation we
assert that the product did not supply any config data for the package.

Bug: 366062990
Change-Id: I209a258c91654515c074b707a09a35b434820a7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118497
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>