[netstack3] No route lookup during route discovery
When installing routes during IPv6 route discovery, skip checking if the
route was already present. Now that the routes are installed in a
dedicated NDP route set, this check is unnecessary.

The check was also inherently racy: Adding routes occurs in bindings,
so it's possible that bindings has a queued request to add the same
route that has not yet been actuated to core.

Bug: 42079625
Change-Id: I6b5c05c2601602c5ff71b490b2970de9494befaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123979
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>