[sestarnix] Eagerly initialize FsNode labeling details
This is a reland of commit 413406bb9941700a7a0fe965e28faf1428ad0f6c,
with a race-condition leading to inconsistent handling of FsNode
labeling during the transition from no-policy to policy-loaded
states addressed.

Previously labels for FsNodes were resolved on-demand, with the
fs_node_effective_sid() helper calling on to the
fs_node_resolve_security_label() helper.

Labels are now resolved for unlabeled FsNodes directly by the
fs_node_init_with_dentry() hook, when they are first linked into
a DirEntry. e.g. for genfscon they will be labeled based on their
FileSystem-relative path, and for fs_use_xattr their security
attribute will be read, to populate a label.

Special cases, such as creation of an FsNode for a new file, may
label the node prior to it being linked into a DirEntry for the
first time, in which case the fs_node_init_with_dentry() hook is
a no-op.

This requires that any FsNodes created in FileSystems prior to the
SELinux policy being loaded are tracked, and proactively labeled
along with the FileSystem. The initial implementation retains a
set of weak references to the relevant DirEntries, to allow node
labeling to take account of the FileSystem-relative path.

Bug: 366405530, 369268313, 367585803, 366405587
Run-All-Tests: true
Change-Id: I66a9a9de35fa5f877d8b9ab7f3254aa969f470dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124732
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>