[components][dictionaries] Simplify dynamic dictionaries implementation
Today, we implement dynamic dictionaries by creating a new dictionary
that _extends_ the DictionaryRef obtained from the Router protocol
served by the program.

This extension is an unnecessary step and carries the limitation that if
the program adds new items to the dictionary after it is routed for the
first time, they won't appear in subsequent routes. This is because the
extension operation takes a 'snapshot' of the source dictionary.

Fixed: 368155371
Change-Id: I4c533fc6cf9aef24c2b7c9e3940cf0fe9a748a2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121832
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>