[dl] Move LinkingSession::TraverseDeps to RuntimeModule
This CL is a verbatim code move only, to make review easier. The
`TraverseDeps` function always operates on the `RuntimeModule` that will
serve as the root of its constructed graph.

This function fills out the runtime module's `module_tree` when dlopen
is called on the module. Currently, it is only called on the root module
when it is loaded. This will eventually also need to be called to fill
out the module tree for a module that was loaded as a dependency of
another module, when dlopen is called on that loaded dependency.

As this will always operate on a root `RuntimeModule`, moving this
method to the `RuntimeModule` class will make it easier to invoke for
both scenarios.

Bug: 354786114
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I0395a7b8a4f1a25b0a36fb8ff21e71047f0ed58f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124680
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>