[kernel][vm] Correct SOLPI assert
Due to how chainlock transactions work and the need to perform backoff
it is possible that a thread sets is_ready_for_waiter_, and yet does
not actually end up blocking. In this case the thread will have
observed that the stack ownership interval has ended, and so there is
no problem with this case, except for this assert spuriously
triggering.

Change-Id: Id4794e37d3f3735f4a9589ba5cf6815d3ebc8bde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115313
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>