[dns/netemul] Split DHCP test into DHCPv4 and NDP
Separate the existing DHCPv4 and NDP test into two separate tests. We
expect the Netstack3 out-of-stack DHCPv4 client to currently update the
list of servers used by the DnsResolver, but we don't expect to see NDP
messages. The existing test was failing but it was unclear whether due
to DHCPv4 servers not existing, NDP servers not existing, or neither
existing.

Create a library for the netcfg integration tests to expose a
`with_netcfg_owned_device` that can be used by other netemul tests to
simulate an environment where a device is discovered in devfs, picked up
by Netcfg, and then observed through the Interfaces watcher. This is
required for the DNS tests involving netcfg to return the correct
results.

Bug: 328647117

Change-Id: Idb62ee4ea28275e56822e984cc29dabe563ef54e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125237
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>